+++ 2019/08/template/heiseirb_api_client/lib/heiseirb_api_client/crud/user.rb
@@ -86,5 +86,30 @@ module HeiseirbApiClient
       end
       response.status
     end
+
+    def similar_users(user_id, limit = 3)
+      event_schedule_ids = fetch_event_schedule_ids_by_target_user(user_id)
+      similar_user_ids = []
+      event_schedule_ids.each do |id|
+        response = @conn.get do |req|
+          req.url "/client/event_schedules/#{id}"
+          req.headers = HEADRS
+        end
+        similar_user_ids.concat(response.body['data']['attributes']['users']['confirmed_user_ids'])
+      end
+      # ↓流石にワンライナーしすぎやろワロタ
+      similar_user_ids.tally.invert.sort {|a, b| b <=> a }.reject {|_key, value| value == user_id }.map {|_key, value| value}.take(limit)
+    end
+
+    private 
+
+    def fetch_event_schedule_ids_by_target_user(user_id)
+      response = @conn.get do |req|
+        req.url "/client/users/#{user_id}"
+        req.headers = HEADRS
+      end
+
+      response.body["data"]["relationships"]["join_schedules"]["data"].map {|es| es["id"].to_i }
+    end
   end
 end
